fix(sidebar): lifecycle cleanup, eager loading, extract sidebar enums
P1: Remove dead .onDisappear{stopMonitoring()} — SidebarView never
disappears in NavigationSplitView. MarketStatusService intentionally
polls for app lifetime. Timer invalidated on deinit.

P1: Add eager paperTradingService.loadPositions() in ContentView.task
so the sidebar green dot populates on app launch, not only after
navigating to Paper Trading.

P2: Extract sidebar enums (ResearchSection, BuildSection, TradeSection,
SidebarSection) to SidebarModels.swift — keeps ContentView under 300
lines per institutional learning.
